refactor(gpt2): into a gpt2 subpackage
