--- toyllm/__init__.py
@@ -1,4 +1,0 @@
-from .model import GPTModel, TextGenerator
-from .tokenizer import gpt2_tokenizer
-
-__all__ = ["GPTModel", "TextGenerator", "gpt2_tokenizer"]
