--- toyllm/model/__init__.py
@@ -1,4 +1,0 @@
-from .generate import TextGenerator
-from .gpt import GPTModel
-
-__all__ = ["GPTModel", "TextGenerator"]
